requirement.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
langchain-openai
fastapi 
uvicorn
pydantic